matching_dog/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
numpy
Pillow
torch
torchvision
open_clip_torch
ultralytics
